Change src/* to read resource type specs from YAML files
This is an intermediary stepes towards making resource types be
specified via the Qvarn API, but before we can work on that, we need
to change things so that the code can get the specs from data and
handles all resource types in one process.

The YAML files contain exactly the same resource type specs that used
to be in Python, but in data files instead of in code. Resource type
specific validation code has been dropped, since there will be no way
to safely allos API users to provide such code; most of this will
eventually be brought back via a stronger type system for resource
types.